feat(vimrc): initial commit of vimrc
This is a template of the vimrc that I plan to make consistent between
my different writing/programming environments. The intent is to make a
python script that will "install" it on new systems and fill in template
sections such as %%%TOP-NOTES-PATH%%% based on where I put folders for
that computer.

Refs: