feat(bashrc): initial commit of bashrc
This is a template of the bashrc that I plan to make consistent between
my different writing/programming environments. The intent is to make a
python script that will "install" it on new systems and fill in template
sctions such as %%%TOP-NOTES-PATH%%% based on where I put folders for
that computer.

Refs: