feat(vimrc): add monthly note creation to vimrc
I've got a script to create monthly notes just like I've got for weekly
and daily notes. This commit adds calling that script and switching to
the new note automatically to the vim config as well as binding it to a
hot key just like for daily and weekly notes. This will make interacting
with the monthly note more smooth and convenient.

Refs: