feat(vimrc): add toggle highlighting to vimrc
Vim has functionality to use a highlight color on search terms. This
is off by default but with a simple command can be turned on. I like
leaving it off most of the time but wanted a quick way to toggle it on.
This commit adds to the vimrc a quick keybind to toggle highlighting.

Refs: