feat(vimrc): change the mapping that runs pytest to go to specific tab
The original mapping for running pytest would move over a single tab
from whereever the mapping was run and attempt to run pytest. This
would cause problems if I had more than two tabs open at a time and "one
tab over" wasn't my terminal tab. This change causes the mapping to
specifically attempt to switch to the second tab and run pytest. This
will be much more consistent since that's the tab I tend to use for my
terminals.

Refs: