feat(vimrc): add custom command to populate local list with pylint
This commit will add a command to run pylint on the file in the current
buffer and output the linters findings to the local list window. It'll
then open the local list window for viewing. This is added to a key bind
and will streamline linting when working in python.

Refs: