feat(vimrc): add keybind to find note tags
This commit adds a keybind that will automatically find what all unique
tags (ex. #refactoring, #source, etc.) exist in a folder. This will make
working with tags one step easier when in vim as they can all now be
seen in one convenient place. Ideally this functionality will be
expanded in the future.

Refs: