feat(Setup Environment): add basic script to create environment variables
This repository makes use of environment variables in its configuration
files (ex. vimrc, bashrc). This commit adds a python script that will go
through and ask a user to provide definition for any of these variables
that do not already exist on a computer so that they're available for
the configuration files.

Refs: