feat(vimrc): change foldmethod to marker for vimscript and add markers in vimrc
This commit changes the foldmethod for vimscript files to be `marker`. It also
adds in a bunch of markers to wrap each section and individual functions. It's
a bit of extra syntax in the file but should make working with it (and other
vimscript files) much easier as I can now fold away sections I'm not actively
working on/looking at.

Refs: