feat(vimrc): add new project and backlog note automation
This commit ties the new project/backlog note creation script to the
vimrc. Also it makes it so that the project main and backlog main notes
are uneditable when opened using their keybindings. This was done so that
only the script populates them as it'd overwrite anything custom added
anyways.

Refs: