fix(vimrc): make productivity main nonmodifiable local
When I created functions to make the project/backlog main notes and open
them, I set modifiable to off. This ended up working on the vim instance
in general though instead of just the productivity main notes like I
wanted. This fix makes the setting local to the productivity main note
buffer and so the other buffers are still modifiable.

Refs: