fix(vimrc): move comment for open note tags keybind
In a previous commit I used the `<keybind>| " comment` hack to put in
comments. When I did it for the note tags keybind, I didn't account for
the line wraps. This commit moves the comment to the end of the line
wraps so that the keybind works as intended again.

Refs: