feat(Fix Note Priorities): add script that fixes backlog note priorites
In my notes I've got a folder of backlog items that I've prioritized by
adding a two digit prefix to the file names. Over time as I pull items
out of the backlog to work on or archive them, the numbering scheme will
get off. This commit introduces a script that will go in and
automatically roll the remaining numbers up so that they're contiguous
from 1 to n where n is the number of prioritized notes in the given
folder.

Refs: