feat(Grep Operator): add a grep operator function/command
This commit adds a grep operator for vim that allows easily searching
words and uses the built in grep which is faster than vimgrep. This will
make searching both faster and easier to do on a regular basis. This
commit replaces the `placeholder.vim` file that I had placed to setup
the package directory structure.

Refs: