refactor(noteplug): setup separate productivity main note options
I've got functionality set up to automatically open productivity main
notes (project main and backlog main) as non-modifiable/read only
buffers using a keybind. This commit moves some of that configuration
from the function that opens the files to autocommands. That way the
non-modifiable will still be in effect even if I open the files up in a
way other than the keybinds. Also, this gives a spot for future
functionality to be added to the notes such as setting them up as custom
menus.

Refs: