chore(Setup Environment): remove setting custom scripts path
It appeared that the custom scripts path environment variable wasn't
working in getting scripts to run in git bash. The solution ended up
being to create a `bin` folder at the home directory and putting files
there since that was already on the system path. This commit removes the
setting of the environment variable in the setup script.

Refs: