chore(vimrc): update vimrc fold markers
I use the vim style fold markers in my vimscript files. I originally
used the open and closed curly brace marker method in my vimrc but it's
harder to maintain because you have to keep track of where the closed
braces are and make sure they line up with the number of open braces.
This commit changes the method to use the `{{{1`, `{{{2` version of
marker folder which should be easier to maintain.

Refs: