feat(vimrc): change insert mode word deletion
This commit changes the word deletion in insert mode to use control I.
The mapping before had been for inserting a tab but that function is
already covered by the tab key itself. By changing where the delete word
key is located, all of the insert mode delete keys are on the same hand
so that there's no more switching back and forth of which hand is
holding control.

Refs: