feat(Vim Python): add docstring section autocompletion
In docstrings it's tedious to type out the whole section header and
dashes for each section. This commit adds autocompletion after typing
the beginning of each section header. Note there were some difficulties
in implementing this as the autoindent puts the cursor where it should
be for the section header while typing but when the command is run the
cursor gets reset to the beginning of the line. That is handled by the
extra character insertion and deletion at the beginning of each mapping.

Refs: