feat(Vim Statusline): add file that controls vim statusline
This commit adds a basic file that will be expanded to improve and
customize the statusline in vim. It sets it to the basic default
features that I currently want. By customizing the statusline I can
remove the bloat that I never use and put in just the features that I
want to use.

Refs: