feat(Vim Statusline): add git branch to the statusline
This commit adds the current branch of a git repository to the
statusline. If the file/buffer is not part of a git repository it will
just show blank. This was implemented through autocommands and buffer
specifc variables because if the function was called directly then vim
would begin to lag significantly. This will make it easier to track what
branch I'm on as I'm editing so that I don't accidentally write code on
the wrong branch.

Refs: