feat(Fix Note Priorities): add script to insert file with new priority
This commit adds a script that will take a file and a priority number
and insert it into a folder of prioritized files (tasks) where the
priority is coded as the padded two digit prefix for the file (ex.
03-task.md). This will make adding new tasks into the middle of a
prioritized set of tasks much easier as there will be no need to
manually adjust the existing files to accommodate the new one.

Refs: