feat(Vim Statusline): update statusline colors to match wombat
Currently vim uses the wombat color scheme overall but I have custom colors in
the statusline. This commit updates the statusline colors to match what the
airline plugin uses for syncing with the wombat color scheme. This should make
the editor easier on the eyes and match better between the buffers and the
statuslines.

Refs: