feat(vimrc): add spell check toggle to vimrc
Vim has a built in spell checker that is off by default. This makes
sense when working with code as a lot of the variables would just
randomly get marked red. There are times though that I'd want to be able
to turn it on such as writing commit messages and docstrings. This
commit adds the keybind "cos" to Change Option Spelling so that I can
toggle it.

Refs: