fix(vimrc): remove delete word keybind
I had added a delete word keybind to vim for CTRL-I so that I could use
all of the different delete word keybinds with one hand (H, U and I) and
not have to switch hands for delete character vs word/line with CTRL-W.
When I switched to using windows terminal, suddenly tab started also
deleting words. It turns out CTRL-I and TAB both send '\t' to the
terminal and so the remap affected tab also. Why it worked for vim in
git-bash and not vim in git-bash in windows terminal I do not know.

Refs: