feat(Vim Python): change PylintLOpen keybind
The linting keybind for python was set to leader-pl for "python lint".
As I add more languages to my toolbelt though I thought it'd be a good
idea to use the same kyebind for all of them. Therefore, I moved the
keybind to just space-l for "lint" and other filetype plugins can use
the same keybind. It's also quicker and easier to hit.

Refs: