feat(vimrc): change open terminal keybind
I added a keybind to make openning a terminal in a new window easy
awhile back. It was using CTRL-` as the bind but that keybind doesn't
work in Windows Terminal. This commit changes the binding to one that
will work in Windows Terminal where I do most of my programming now.

Refs: