feat(vimrc): disable option that causes windows resize on close
When I've got multiple windows open in vim, I've sometimes got a couple
that are "minimized". When I close an unrelated window vim tends to
cause all of the windows to resize back to default equal splits. It
turns out this behavior is caused by the option `equalalways`. This
commit disables that option.

Refs: