feat(vimrc): force the vim status line to default to on
Vim has an option `laststatus` that determines if the status line should
default to off, only show up when there are multiple windows, or just
always be on. This commit sets the option to have the status line always
be on so that my environment is consistent system to system.

Refs: