feat(Fuzzy Search): add ignore folders to open dialog
This commit adds the .venv and build folders to the ignore list for my
open dialog. Those folders are populated by python virtual environment
stuff or the outputs of compiling and building. Neither of which are
files that I want to search for with my open dialog and they both end up
with LOTS of files diluting the usefulness of my dialog.

Refs: