feat(i3): set desktop background automatically on start up
This commit adds a line to the i3 config to run feh automatically on
start up to set the desktop background. Note, the path to the wallpaper
is currently hardcoded. If a different file is wanted in the future,
it'd have to be changed directly in the config. There are probably more
elegant solutions but this one worked for now.

Refs: